requirements.txt
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5       
chardet==4.0.0
ChatterBot==1.0.8        
click==7.1.2
cymem==2.0.5
en-core-web-sm==3.0.0    
Flask==1.1.2
idna==2.10
importlib-metadata==3.7.2
itsdangerous==1.1.0      
jellyfish==0.8.2
Jinja2==2.11.3
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.5
mysql-connector-python==8.0.23
numpy==1.20.1
packaging==20.9
pathy==0.4.0
preshed==3.0.5
protobuf==3.15.6
pt-core-news-sm==3.0.0
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.1
SQLAlchemy==1.3.23
srsly==2.4.0
thinc==8.0.2
tqdm==4.59.0
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
zipp==3.4.1